refactor(server): improve server lifecycle and autocert management
- Separate server Start() and Run() methods for better error handling
- Provide errgroup-compatible Run() method
- Update context cancellation handling
- Add separate Stop() method replacing Shutdown()
- Remove direct logging with context
- Precompile HTML templates for provisioning and error pages
- Centralize constants like DefaultDomainLookupTimeout
- Add server shutdown timeout configuration option
- Enhance server/autocert error handling and logging
- Improve context and error handling in server lifecycle
