+++ core/server/autocert.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"crypto/tls"
 	"errors"
-	"html/template"
 	"io"
 	"log/slog"
 	"net/http"
@@ -99,15 +98,17 @@ type AutoCertServer struct {
 	notFoundHandler     NotFoundHandler
 
 	// State
-	running bool
+	running         bool
+	shutdownTimeout time.Duration
 }
 
 // NewAutoCertServer creates a new server with automatic HTTPS support.
 func NewAutoCertServer(opts ...AutoCertOption) (*AutoCertServer, error) {
 	// Default configuration
 	s := &AutoCertServer{
-		httpServer:  New(":80"),  // Default HTTP address
-		httpsServer: New(":443"), // Default HTTPS address
+		httpServer:      New(":80"),       // Default HTTP address
+		httpsServer:     New(":443"),      // Default HTTPS address
+		shutdownTimeout: 30 * time.Second, // Default shutdown timeout
 	}
 
 	for _, opt := range opts {
@@ -184,191 +185,11 @@ func DefaultNotFoundHandler(logger *slog.Logger) NotFoundHandler {
 	}
 }
 
-// buildProvisioningHTML generates the provisioning status page
-func buildProvisioningHTML(info *DomainInfo) string {
-	tmpl := `<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <meta http-equiv="refresh" content="10">
-    <title>Securing Connection</title>
-    <style>
-        body {
-            font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif;
-            background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-            min-height: 100vh;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            margin: 0;
-            padding: 20px;
-        }
-        .container {
-            background: white;
-            border-radius: 12px;
-            box-shadow: 0 20px 60px rgba(0,0,0,0.15);
-            padding: 40px;
-            max-width: 500px;
-            width: 100%;
-            text-align: center;
-        }
-        h1 { color: #333; margin: 0 0 20px; font-size: 28px; }
-        .lock-icon { font-size: 48px; margin-bottom: 20px; }
-        .domain { color: #667eea; font-weight: 600; font-size: 18px; }
-        p { color: #666; line-height: 1.6; margin: 15px 0; }
-        .progress {
-            width: 100%;
-            height: 6px;
-            background: #f0f0f0;
-            border-radius: 3px;
-            overflow: hidden;
-            margin: 30px 0;
-        }
-        .progress-bar {
-            height: 100%;
-            background: linear-gradient(90deg, #667eea, #764ba2);
-            animation: progress 2s ease-in-out infinite;
-        }
-        @keyframes progress {
-            0% { transform: translateX(-100%); }
-            50% { transform: translateX(0); }
-            100% { transform: translateX(100%); }
-        }
-    </style>
-</head>
-<body>
-    <div class="container">
-        <div class="lock-icon">🔒</div>
-        <h1>Setting up secure connection</h1>
-        <p>We're configuring SSL/TLS for</p>
-        <p class="domain">{{.Domain}}</p>
-        <div class="progress"><div class="progress-bar"></div></div>
-        <p>This typically takes 30-60 seconds.</p>
-        <p style="font-size: 14px; color: #999;">This page will refresh automatically</p>
-    </div>
-</body>
-</html>`
-
-	result := strings.ReplaceAll(tmpl, "{{.Domain}}", template.HTMLEscapeString(info.Domain))
-	return result
-}
-
-func buildFailedHTML(info *DomainInfo) string {
-	tmpl := `<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Configuration Required</title>
-    <style>
-        body {
-            font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif;
-            background: #f5f5f5;
-            min-height: 100vh;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            margin: 0;
-            padding: 20px;
-        }
-        .container {
-            background: white;
-            border-radius: 12px;
-            box-shadow: 0 4px 20px rgba(0,0,0,0.1);
-            padding: 40px;
-            max-width: 600px;
-            width: 100%;
-        }
-        h1 { color: #d93025; margin: 0 0 20px; font-size: 28px; }
-        .domain {
-            color: #333;
-            font-weight: 600;
-            background: #f8f9fa;
-            padding: 8px 12px;
-            border-radius: 6px;
-            display: inline-block;
-            margin: 10px 0;
-        }
-        .error-box {
-            background: #fef2f2;
-            border: 1px solid #fecaca;
-            border-radius: 8px;
-            padding: 16px;
-            margin: 20px 0;
-        }
-        .error-message {
-            color: #b91c1c;
-            font-family: 'Courier New', monospace;
-            font-size: 14px;
-            word-break: break-all;
-        }
-        h3 { color: #333; margin: 30px 0 15px; }
-        ul { color: #666; line-height: 1.8; }
-        li { margin: 8px 0; }
-    </style>
-</head>
-<body>
-    <div class="container">
-        <h1>⚠️ Domain Configuration Required</h1>
-        <p>Unable to set up SSL/TLS certificate for:</p>
-        <div class="domain">{{.Domain}}</div>
-        <div class="error-box">
-            <strong>Error Details:</strong>
-            <div class="error-message">{{.Error}}</div>
-        </div>
-        <h3>Common causes:</h3>
-        <ul>
-            <li>DNS records not properly configured</li>
-            <li>Domain not pointing to our servers</li>
-            <li>CAA records blocking Let's Encrypt</li>
-            <li>Rate limits exceeded</li>
-        </ul>
-        <p>Please verify your DNS settings and contact support if the issue persists.</p>
-    </div>
-</body>
-</html>`
-
-	result := strings.ReplaceAll(tmpl, "{{.Domain}}", template.HTMLEscapeString(info.Domain))
-	result = strings.ReplaceAll(result, "{{.Error}}", template.HTMLEscapeString(info.Error))
-	return result
-}
-
-const defaultNotFoundHTML = `<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>404 - Domain Not Found</title>
-    <style>
-        body {
-            font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif;
-            background: #f5f5f5;
-            min-height: 100vh;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            margin: 0;
-        }
-        .container { text-align: center; }
-        h1 { font-size: 120px; color: #e0e0e0; margin: 0; font-weight: 700; }
-        h2 { color: #333; margin: 20px 0; font-size: 28px; }
-        p { color: #666; font-size: 18px; }
-    </style>
-</head>
-<body>
-    <div class="container">
-        <h1>404</h1>
-        <h2>Domain Not Found</h2>
-        <p>The requested domain is not configured on this server.</p>
-    </div>
-</body>
-</html>`
-
-// Run starts both HTTP and HTTPS servers with the provided handler.
+// Start starts both HTTP and HTTPS servers with the provided handler.
 // The HTTP server handles ACME challenges and redirects to HTTPS.
 // The HTTPS server serves the application with TLS.
-func (s *AutoCertServer) Run(ctx context.Context, handler http.Handler) error {
+// This is a blocking operation that returns when the context is canceled or an error occurs.
+func (s *AutoCertServer) Start(ctx context.Context, handler http.Handler) error {
 	s.mu.Lock()
 	if s.running {
 		s.mu.Unlock()
@@ -385,26 +206,26 @@ func (s *AutoCertServer) Run(ctx context.Context, handler http.Handler) error {
 
 	errCh := make(chan error, 2)
 
-	// Run HTTP server with ACME handler
+	// Start HTTP server with ACME handler
 	go func() {
-		if err := s.httpServer.Run(ctx, s.createHTTPHandler()); err != nil {
+		if err := s.httpServer.Start(ctx, s.createHTTPHandler()); err != nil {
 			errCh <- errors.Join(ErrHTTPServer, err)
 		}
 	}()
 
-	// Run HTTPS server with application handler
+	// Start HTTPS server with application handler
 	go func() {
-		if err := s.httpsServer.Run(ctx, handler); err != nil {
+		if err := s.httpsServer.Start(ctx, handler); err != nil {
 			errCh <- errors.Join(ErrHTTPSServer, err)
 		}
 	}()
 
 	select {
 	case err := <-errCh:
-		_ = s.Shutdown(ctx)
+		_ = s.Stop()
 		return err
 	case <-ctx.Done():
-		return s.Shutdown(ctx)
+		return ctx.Err()
 	}
 }
 
@@ -462,7 +283,7 @@ func (s *AutoCertServer) getCertificate(hello *tls.ClientHelloInfo) (*tls.Certif
 	}
 
 	// Use timeout to prevent slow domain lookups from blocking TLS handshake
-	ctx, cancel := context.WithTimeout(hello.Context(), 10*time.Second)
+	ctx, cancel := context.WithTimeout(hello.Context(), DefaultDomainLookupTimeout)
 	defer cancel()
 	info, err := s.domainStore.GetDomain(ctx, domain)
 	if err != nil {
@@ -475,8 +296,8 @@ func (s *AutoCertServer) getCertificate(hello *tls.ClientHelloInfo) (*tls.Certif
 	return s.certManager.GetCertificate(hello)
 }
 
-// Shutdown gracefully shuts down both HTTP and HTTPS servers.
-func (s *AutoCertServer) Shutdown(ctx context.Context) error {
+// Stop gracefully shuts down both HTTP and HTTPS servers.
+func (s *AutoCertServer) Stop() error {
 	s.mu.Lock()
 	defer s.mu.Unlock()
 
@@ -487,11 +308,11 @@ func (s *AutoCertServer) Shutdown(ctx context.Context) error {
 	var httpErr, httpsErr error
 
 	if s.httpServer != nil {
-		httpErr = s.httpServer.Shutdown(ctx)
+		httpErr = s.httpServer.Stop()
 	}
 
 	if s.httpsServer != nil {
-		httpsErr = s.httpsServer.Shutdown(ctx)
+		httpsErr = s.httpsServer.Stop()
 	}
 
 	s.running = false
@@ -501,4 +322,30 @@ func (s *AutoCertServer) Shutdown(ctx context.Context) error {
 	}
 
 	return nil
+}
+
+// Run provides errgroup compatibility for coordinated lifecycle management.
+// Returns a function that starts the server, monitors context cancellation,
+// and performs graceful shutdown when the context is cancelled.
+func (s *AutoCertServer) Run(ctx context.Context, handler http.Handler) func() error {
+	return func() error {
+		errCh := make(chan error, 1)
+		go func() {
+			errCh <- s.Start(ctx, handler)
+		}()
+
+		select {
+		case <-ctx.Done():
+			if stopErr := s.Stop(); stopErr != nil {
+				s.httpServer.logger.Error("failed to stop autocert server during context cancellation", "error", stopErr)
+			}
+			<-errCh
+			return nil
+		case err := <-errCh:
+			if errors.Is(err, context.Canceled) || errors.Is(err, context.DeadlineExceeded) {
+				return nil
+			}
+			return err
+		}
+	}
 }
