+++ core/server/autocert_config_test.go
@@ -4,9 +4,10 @@ import (
 	"testing"
 	"time"
 
-	"github.com/dmitrymomot/foundation/core/server"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
+	"github.com/dmitrymomot/foundation/core/server"
 )
 
 func TestNewAutoCertFromConfig(t *testing.T) {
